+++ examples/selectable-rows/index.js
@@ -7,32 +7,32 @@ class Example extends React.Component {
 
     render() {
         const columns = [
-            {   
-                name: 'Name', 
+            {
+                name: 'Name',
                 field: 'name',
                 options: {
                     width: 70,
-                    
+
                 },
             },
             {
-                name: 'Title', 
+                name: 'Title',
                 field: 'title'
             },
             {
-                name: 'Location', 
+                name: 'Location',
                 field: 'location'
             },
             {
-                name: 'Age', 
+                name: 'Age',
                 field: 'age'
             },
             {
-                name: 'Salary', 
+                name: 'Salary',
                 field: 'salary'
             },
             {
-                name: 'Veeeeery loooooooooong title', 
+                name: 'Veeeeery loooooooooong title',
                 field: 'salary',
                 options: {
                     width: 100,
@@ -45,7 +45,7 @@ class Example extends React.Component {
                         );
                     },
                     customValue: (dataObject) => {
-                        return  dataObject.salary +"Edit";
+                        return dataObject.salary + "Edit";
                     },
                 },
             }
@@ -60,7 +60,7 @@ class Example extends React.Component {
                         return (
                             <div>
                                 <Button size="large" variant={"contained"} color={"secondary"}>
-                                    {dataObject.salary * 10} 
+                                    {dataObject.salary * 10}
                                 </Button>
                             </div>
                         );
@@ -71,40 +71,236 @@ class Example extends React.Component {
                     customValue: (dataObject) => {
                         return (dataObject.salary * 10);
                     },
-                    
+
                 },
             }
         ];
 
         const data = [
-            {name: "Veeeeery loooooooooong naaaaaaaaaaame", title: "Title 1", location: "Location 1", age: 30, salary: 10, subFieldData: {value:"11111111"}},
-            {name: "Name 2", title: "Title 2", location: "Location 2", age: 31, salary: 11, subFieldData: {value:"222222222"}},
-            {name: "Name 2", title: "Title 3", location: "Location 2", age: 31, salary: 11, subFieldData: {value:"222222222"}},
-            {name: "Name 2", title: "Title 4", location: "Location 2", age: 31, salary: 11, subFieldData: {value:"222222222"}},
-            {name: "Name 2", title: "Title 5", location: "Location 2", age: 31, salary: 11, subFieldData: {value:"222222222"}},
-            {name: "Name 2", title: "Title 6", location: "Location 2", age: 31, salary: 11, subFieldData: {value:"222222222"}},
-            {name: "Name 2", title: "Title 7", location: "Location 2", age: 31, salary: 11, subFieldData: {value:"222222222"}},
-            {name: "Name 2", title: "Title 8", location: "Location 2", age: 31, salary: 11, subFieldData: {value:"222222222"}},
-            {name: "Name 2", title: "Title 9", location: "Location 2", age: 31, salary: 11, subFieldData: {value:"222222222"}},
-            {name: "Name 2", title: "Title 10", location: "Location 2", age: 31, salary: 11, subFieldData: {value:"222222222"}},
-            {name: "Name 2", title: "Title 11", location: "Location 2", age: 31, salary: 11, subFieldData: {value:"222222222"}},
-            {name: "Name 2", title: "Title 12", location: "Location 2", age: 31, salary: 11, subFieldData: {value:"222222222"}},
-            {name: "Name 2", title: "Title 13", location: "Location 2", age: 31, salary: 11, subFieldData: {value:"222222222"}},
-            {name: "Name 2", title: "Title 14", location: "Location 2", age: 31, salary: 11, subFieldData: {value:"222222222"}},
-            {name: "Name 2", title: "Title 15", location: "Location 2", age: 31, salary: 11, subFieldData: {value:"222222222"}},
-            {name: "Name 2", title: "Title 16", location: "Location 2", age: 31, salary: 11, subFieldData: {value:"222222222"}},
-            {name: "Name 2", title: "Title 17", location: "Location 2", age: 31, salary: 11, subFieldData: {value:"222222222"}},
-            {name: "Name 2", title: "Title 18", location: "Location 2", age: 31, salary: 11, subFieldData: {value:"222222222"}},
-            {name: "Name 2", title: "Title 19", location: "Location 2", age: 31, salary: 11, subFieldData: {value:"222222222"}},
-            {name: "Name 2", title: "Title 20", location: "Location 2", age: 31, salary: 11, subFieldData: {value:"222222222"}},
-            {name: "Name 2", title: "Title 21", location: "Location 2", age: 31, salary: 11, subFieldData: {value:"222222222"}},
-            {name: "Name 2", title: "Title 22", location: "Location 2", age: 31, salary: 11, subFieldData: {value:"222222222"}},
-            {name: "Name 2", title: "Title 23", location: "Location 2", age: 31, salary: 11, subFieldData: {value:"222222222"}},
-            {name: "Name 2", title: "Title 24", location: "Location 2", age: 31, salary: 11, subFieldData: {value:"222222222"}},
-            {name: "Name 2", title: "Title 25", location: "Location 2", age: 31, salary: 11, subFieldData: {value:"222222222"}},
-            {name: "Name 2", title: "Title 26", location: "Location 2", age: 31, salary: 11, subFieldData: {value:"222222222"}},
-            {name: "Name 2", title: "Title 27", location: "Location 2", age: 31, salary: 11, subFieldData: {value:"222222222"}},
-            {name: "Name 2", title: "Title 28", location: "Location 2", age: 31, salary: 11, subFieldData: {value:"222222222"}},
+            {
+                name: "Veeeeery loooooooooong naaaaaaaaaaame",
+                title: "Title 1",
+                location: "Location 1",
+                age: 30,
+                salary: 10,
+                subFieldData: {value: "11111111"}
+            },
+            {
+                name: "Name 2",
+                title: "Title 2",
+                location: "Location 2",
+                age: 31,
+                salary: 11,
+                subFieldData: {value: "222222222"}
+            },
+            {
+                name: "Name 2",
+                title: "Title 3",
+                location: "Location 2",
+                age: 31,
+                salary: 11,
+                subFieldData: {value: "222222222"}
+            },
+            {
+                name: "Name 2",
+                title: "Title 4",
+                location: "Location 2",
+                age: 31,
+                salary: 11,
+                subFieldData: {value: "222222222"}
+            },
+            {
+                name: "Name 2",
+                title: "Title 5",
+                location: "Location 2",
+                age: 31,
+                salary: 11,
+                subFieldData: {value: "222222222"}
+            },
+            {
+                name: "Name 2",
+                title: "Title 6",
+                location: "Location 2",
+                age: 31,
+                salary: 11,
+                subFieldData: {value: "222222222"}
+            },
+            {
+                name: "Name 2",
+                title: "Title 7",
+                location: "Location 2",
+                age: 31,
+                salary: 11,
+                subFieldData: {value: "222222222"}
+            },
+            {
+                name: "Name 2",
+                title: "Title 8",
+                location: "Location 2",
+                age: 31,
+                salary: 11,
+                subFieldData: {value: "222222222"}
+            },
+            {
+                name: "Name 2",
+                title: "Title 9",
+                location: "Location 2",
+                age: 31,
+                salary: 11,
+                subFieldData: {value: "222222222"}
+            },
+            {
+                name: "Name 2",
+                title: "Title 10",
+                location: "Location 2",
+                age: 31,
+                salary: 11,
+                subFieldData: {value: "222222222"}
+            },
+            {
+                name: "Name 2",
+                title: "Title 11",
+                location: "Location 2",
+                age: 31,
+                salary: 11,
+                subFieldData: {value: "222222222"}
+            },
+            {
+                name: "Name 2",
+                title: "Title 12",
+                location: "Location 2",
+                age: 31,
+                salary: 11,
+                subFieldData: {value: "222222222"}
+            },
+            {
+                name: "Name 2",
+                title: "Title 13",
+                location: "Location 2",
+                age: 31,
+                salary: 11,
+                subFieldData: {value: "222222222"}
+            },
+            {
+                name: "Name 2",
+                title: "Title 14",
+                location: "Location 2",
+                age: 31,
+                salary: 11,
+                subFieldData: {value: "222222222"}
+            },
+            {
+                name: "Name 2",
+                title: "Title 15",
+                location: "Location 2",
+                age: 31,
+                salary: 11,
+                subFieldData: {value: "222222222"}
+            },
+            {
+                name: "Name 2",
+                title: "Title 16",
+                location: "Location 2",
+                age: 31,
+                salary: 11,
+                subFieldData: {value: "222222222"}
+            },
+            {
+                name: "Name 2",
+                title: "Title 17",
+                location: "Location 2",
+                age: 31,
+                salary: 11,
+                subFieldData: {value: "222222222"}
+            },
+            {
+                name: "Name 2",
+                title: "Title 18",
+                location: "Location 2",
+                age: 31,
+                salary: 11,
+                subFieldData: {value: "222222222"}
+            },
+            {
+                name: "Name 2",
+                title: "Title 19",
+                location: "Location 2",
+                age: 31,
+                salary: 11,
+                subFieldData: {value: "222222222"}
+            },
+            {
+                name: "Name 2",
+                title: "Title 20",
+                location: "Location 2",
+                age: 31,
+                salary: 11,
+                subFieldData: {value: "222222222"}
+            },
+            {
+                name: "Name 2",
+                title: "Title 21",
+                location: "Location 2",
+                age: 31,
+                salary: 11,
+                subFieldData: {value: "222222222"}
+            },
+            {
+                name: "Name 2",
+                title: "Title 22",
+                location: "Location 2",
+                age: 31,
+                salary: 11,
+                subFieldData: {value: "222222222"}
+            },
+            {
+                name: "Name 2",
+                title: "Title 23",
+                location: "Location 2",
+                age: 31,
+                salary: 11,
+                subFieldData: {value: "222222222"}
+            },
+            {
+                name: "Name 2",
+                title: "Title 24",
+                location: "Location 2",
+                age: 31,
+                salary: 11,
+                subFieldData: {value: "222222222"}
+            },
+            {
+                name: "Name 2",
+                title: "Title 25",
+                location: "Location 2",
+                age: 31,
+                salary: 11,
+                subFieldData: {value: "222222222"}
+            },
+            {
+                name: "Name 2",
+                title: "Title 26",
+                location: "Location 2",
+                age: 31,
+                salary: 11,
+                subFieldData: {value: "222222222"}
+            },
+            {
+                name: "Name 2",
+                title: "Title 27",
+                location: "Location 2",
+                age: 31,
+                salary: 11,
+                subFieldData: {value: "222222222"}
+            },
+            {
+                name: "Name 2",
+                title: "Title 28",
+                location: "Location 2",
+                age: 31,
+                salary: 11,
+                subFieldData: {value: "222222222"}
+            },
         ];
 
         const options = {
@@ -114,12 +310,17 @@ class Example extends React.Component {
             filterType: 'multiselect',
             responsive: 'stacked',
             rowsPerPage: 10,
+            searchText: "Title 20",
+            page: 0,
+            filterList:[[],[],["Location 2"],[],[],[],[]],
+            
             onRowsSelect: (rowsSelected, allRows) => {
                 console.log(rowsSelected, allRows);
             },
             onRowsDelete: (rowsDeleted) => {
                 console.log(rowsDeleted, "were deleted!");
             },
+
             onChangePage: (numberRows) => {
                 console.log(numberRows);
             },
@@ -140,6 +341,9 @@ class Example extends React.Component {
             },
             onRowClick: (rowData, rowState) => {
                 console.log(rowData, rowState);
+            },
+            onTableChange: (action, state) => {
+                console.log(action, state);
             }
         };