+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+<a name="0.2.6"></a>
+## 0.2.6 (2018-12-25)
+
+* Added option searchText?: string
+* Added event onTableChange?: (action: string, tableState: MaterialDatatableState) to TS definition
+* Renamed MuiDatatablesTableState to MaterialDatatableState in TS definition
+
 <a name="0.2.5"></a>
 ## 0.2.5 (2018-12-24)